Refactor pan-types.h into discrete header files for each pan-view code module
