+++ src/pan-timeline.h
@@ -19,21 +19,12 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-#ifndef PAN_VIEW_H
-#define PAN_VIEW_H
+#ifndef PAN_TIMELINE_H
+#define PAN_TIMELINE_H
 
 #include "main.h"
 #include "pan-types.h"
 
-void pan_window_new(FileData *dir_fd);
-
-GList *pan_layout_intersect(PanWindow *pw, gint x, gint y, gint width, gint height);
-void pan_layout_resize(PanWindow *pw);
-
-void pan_cache_sync_date(PanWindow *pw, GList *list);
-
-GList *pan_cache_sort(GList *list, SortType method, gboolean ascend);
-
+void pan_timeline_compute(PanWindow *pw, FileData *dir_fd, gint *width, gint *height);
 
 #endif
-/* vim: set shiftwidth=8 softtabstop=0 cindent cinoptions={1s: */
