+++ src/pan-view.c
@@ -19,23 +19,31 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-#include "main.h"
 #include "pan-view.h"
 
 #include "bar_exif.h"
 #include "dnd.h"
 #include "editors.h"
 #include "exif.h"
-#include "metadata.h"
 #include "fullscreen.h"
 #include "history_list.h"
+#include "image.h"
 #include "img-view.h"
 #include "menu.h"
+#include "metadata.h"
 #include "misc.h"
-#include "pan-types.h"
+#include "pan-calendar.h"
+#include "pan-folder.h"
+#include "pan-grid.h"
+#include "pan-item.h"
+#include "pan-timeline.h"
+#include "pan-util.h"
+#include "pixbuf-renderer.h"
+#include "pixbuf_util.h"
 #include "thumb.h"
 #include "ui_fileops.h"
 #include "ui_menu.h"
+#include "ui_misc.h"
 #include "ui_tabcomp.h"
 #include "ui_utildlg.h"
 #include "uri_utils.h"
